backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
python-multipart>=0.0.9
pydantic>=2.7.4

pydantic-settings>=2.3.4
yfinance>=0.2.40
requests>=2.32.3
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
slowapi>=0.1.9
fastapi-mail>=1.4.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
cryptography>=42.0.5
web3>=6.15.1

